perf(magic-vars): materialize $*DISTRO/$*PERL/$*RAKU/$*VM/$*KERNEL lazily on first read
Slice 2 of todo/tickets/magic-vars-should-be-built-lazily.md. These five
Instances (Version parses, a 32-element signal array, the VM class's
vm_config hash) were built eagerly at every Interpreter::new()/thread-clone
and hoisted into the base tier, even for programs that never read them.
Profiling (strace -c + perf stat) confirmed construction, not syscalls,
dominated the eager cost on Linux (slice 1 already reduced syscalls to one
cached uname(2)): ~11% fewer instructions retired for `mutsu -e 'say 1'`.

They now materialize on first read via Interpreter::lazy_magic_dynamic_var,
called from the dynamic-var read miss path (get_env_with_main_alias_inner),
still cached process-wide through the existing per-var OnceLocks.
make_perl_instance's version attribute now reads current_language_version()
directly at construction time so a lazily-built $*RAKU/$*PERL still reflects
the compile unit's `use v6.x`.
